// src/dtos/users-database.dto.ts
export interface UserDatabaseRow {
  id: string | number;
  name?: string;
  nome?: string;
  lastName?: string;
  sobrenome?: string;
  quantity?: number;
  quantidade?: number;
  type?: string;
  tipo?: string;
  email: string;
  createdDate?: string;
  dataCriada?: string;
  createdTime?: string;
  horaCriada?: string;
  updatedDate?: string;
  dataAtualizada?: string;
  updatedTime?: string;
  horaAtualizada?: string;
}